Add vuex
Added the ability to include tweets via vuex.
The router has been configured for each element of the navigation.
An api.js file was created where an array of elements can be located, they are displayed dynamically on the screen.
The connection with the data api was commented for an ease when creating the new functions